+++ include/core/sampler/SeedSampler.hpp
@@ -62,7 +62,7 @@ template <unsigned int dim>
 Vector<dim> SeedSampler<dim>::getSample() {
     Vector<dim> vec;
     for (unsigned int i = 0; i < dim; ++i) {
-        vec[i] = this->m_minBoundary[i] + (float)(m_randomEngine() % (int)(this->m_maxBoundary[i] - this->m_minBoundary[i]));
+        vec[i] = this->m_minBoundary[i] + (double)(m_randomEngine() % (int)(this->m_maxBoundary[i] - this->m_minBoundary[i]));
     }
     return vec;
 }
